main story page, subscription, like

+++ src/main/java/com/cos/photogramstart/web/dto/subscribe/SubscribeDto.java
@@ -2,14 +2,16 @@ package com.cos.photogramstart.web.dto.subscribe;
 
 import lombok.*;
 
+import java.math.BigInteger;
+
 @Builder
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class SubscribeDto {
-    private long id; // toUserId로 이용
+    private BigInteger id; // toUserId로 이용
     private String username;
     private String profileImageUrl;
-    private int subscribeState; // 구독한 상태인지
-    private int equalUserState; // pageUserId가 구독중인 유저가 principal과 동일인인지
+    private Integer subscribeState; // 구독한 상태인지
+    private Integer equalUserState; // pageUserId가 구독중인 유저가 principal과 동일인인지
 }
